requirements.txt
# Core Dependencies
numpy>=1.19.0
pandas>=1.2.0
scikit-learn>=0.24.0
sentence-transformers>=2.0.0
faiss-cpu>=1.7.0
pyyaml>=5.4.1
networkx>=2.5
phonenumbers>=8.12.0
python-Levenshtein>=0.12.0
jellyfish>=0.8.0
recordlinkage>=0.14

# Cloud Integration
databricks-connect>=7.3
snowflake-snowpark-python>=0.7.0
azure-synapse-spark>=0.7.0

# API & Web
fastapi>=0.68.0
uvicorn>=0.15.0

# Machine Learning
torch>=1.9.0
transformers>=4.5.0

# Distributed Computing
dask>=2021.6.0

# Data Quality
great-expectations>=0.13.0

# Development Dependencies
pytest>=6.0.0
pytest-cov>=2.12.0
black>=21.5b2
isort>=5.9.0
flake8>=3.9.0
mypy>=0.910
sphinx>=4.0.0
sphinx-rtd-theme>=0.5.0
